rename kv_parse_value to kv_get_value
After exploring this code a bit more I figured out that this is actually a pretty inefficient way to parse kv files. Especially when reading lots of different kv files.

Might make a seperate repo that focus on dealing with lots of kv files.

This still have value for simpler projects where we are scanning for only a few values in a small projects... to keep complexity level down. But in the meantime I've updated the readme and updated the function name to reflect this fact.